added a readfile

+++ src/JointRadiometicCalib/JRC.hpp
@@ -36,13 +36,13 @@ float h_[3][1024]; //derivative for h
 //float grad_log_inv_RF[];
 
 //for kalman filter estimation
-Eigen::MatrixXf D;
-Eigen::MatrixXf b;
-Eigen::MatrixXf u_;
-Eigen::MatrixXf R;
-Eigen::MatrixXf P_new;
-Eigen::MatrixXf c_new;
-Eigen::MatrixXf c_prev;
+Eigen::MatrixXd D;
+Eigen::MatrixXd b;
+Eigen::MatrixXd u_;
+Eigen::MatrixXd R;
+Eigen::MatrixXd P_new;
+Eigen::MatrixXd c_new;
+Eigen::MatrixXd c_prev;
 
 
 bool trackingMode; //0 : knownRF, 1: unknownRF
@@ -52,8 +52,8 @@ JointRadiometicCalib(int nfeatures){
     this->tracker.writeInternalImages = false;
     this->tracker.affineConsistencyCheck = -1;  /* set this to 2 to turn on affine consistency check */
     this->nfeatures = nfeatures;
-    this->c_new = Eigen::MatrixXf::Zero(4,1);
-    this->P_new = Eigen::MatrixXf::Identity(4,4);
+    this->c_new = Eigen::MatrixXd::Zero(4,1);
+    this->P_new = Eigen::MatrixXd::Identity(4,4);
     
 }
 ~JointRadiometicCalib(){}
@@ -64,23 +64,23 @@ void setRFderivatives();
 float getRF_Value(int idx);
 float getRF_DerivValue(int idx);
 
-Eigen::MatrixXf get_U(int window_size, int x, int y, Mat a,Mat b);
-Eigen::MatrixXf get_U(int window_size, int x, int y, Mat aM,Mat bM,Mat pM[],Mat qM[]);
-Eigen::MatrixXf get_w(int window_size, int x, int y, Mat aM, Mat bM);
-Eigen::MatrixXf get_w(int window_size, int x, int y, Mat aM, Mat bM,Mat rM[],Mat pM[],Mat qM[]);
-Eigen::VectorXf get_v(int window_size, int x, int y, Mat aM, Mat bM, Mat betaM);
-Eigen::VectorXf get_v(int window_size, int x, int y, Mat aM, Mat bM, Mat dM, Mat pM[], Mat qM[]);
-Eigen::VectorXf get_z(bool JRCtrackingMode);
-Eigen::MatrixXf get_lamda(int window_size,int x, int y);
-Eigen::MatrixXf get_lamda(int window_size,int x, int y, Mat rM[]);
-Eigen::VectorXf get_m(int window_size, int x, int y, Mat betaM);
-Eigen::VectorXf get_m(int window_size, int x, int y, Mat dM, Mat rM[]);
+Eigen::MatrixXd get_U(int window_size, int x, int y, Mat a,Mat b);
+Eigen::MatrixXd get_U(int window_size, int x, int y, Mat aM,Mat bM,Mat pM[],Mat qM[]);
+Eigen::MatrixXd get_w(int window_size, int x, int y, Mat aM, Mat bM);
+Eigen::MatrixXd get_w(int window_size, int x, int y, Mat aM, Mat bM,Mat rM[],Mat pM[],Mat qM[]);
+Eigen::VectorXd get_v(int window_size, int x, int y, Mat aM, Mat bM, Mat betaM);
+Eigen::VectorXd get_v(int window_size, int x, int y, Mat aM, Mat bM, Mat dM, Mat pM[], Mat qM[]);
+Eigen::VectorXd get_z(bool JRCtrackingMode);
+Eigen::MatrixXd get_lamda(int window_size,int x, int y);
+Eigen::MatrixXd get_lamda(int window_size,int x, int y, Mat rM[]);
+Eigen::VectorXd get_m(int window_size, int x, int y, Mat betaM);
+Eigen::VectorXd get_m(int window_size, int x, int y, Mat dM, Mat rM[]);
 
-void blockAllMatrix(int numOfTrackFeature,Eigen::MatrixXf &Uinv_all,Eigen::MatrixXf &w_all,Eigen::VectorXf &v_all,Eigen::MatrixXf &lamda_all,Eigen::VectorXf &m_all, bool JRCtrackingMode);
-float get_K(kltFeature f,Eigen::MatrixXf Uinv_all,Eigen::MatrixXf w_all,Eigen::VectorXf v_all,Eigen::MatrixXf lamda_all,Eigen::VectorXf m_all,bool JRCtrackingMode);
+void blockAllMatrix(int numOfTrackFeature,Eigen::MatrixXd &Uinv_all,Eigen::MatrixXd &w_all,Eigen::VectorXd &v_all,Eigen::MatrixXd &lamda_all,Eigen::VectorXd &m_all, bool JRCtrackingMode);
+float get_K(kltFeature f,Eigen::MatrixXd Uinv_all,Eigen::MatrixXd w_all,Eigen::VectorXd v_all,Eigen::MatrixXd lamda_all,Eigen::VectorXd m_all,bool JRCtrackingMode);
 void initialization(kltFeature f);
 void constructMatrix(kltFeature f,int pylevel,int window_size, int x, int y, Mat J_origin, Mat I_origin,Mat J_gradx, Mat I_gradx,Mat J_grady, Mat I_grady, bool JRCtrackingMode);
-void constructAllMatrix(kltFeature f, int numOfTrackFeature, Eigen::MatrixXf &Uinv_all,Eigen::MatrixXf &w_all,Eigen::VectorXf &v_all,Eigen::MatrixXf &lamda_all,Eigen::VectorXf &m_all);
+void constructAllMatrix(kltFeature f, int numOfTrackFeature, Eigen::MatrixXd &Uinv_all,Eigen::MatrixXd &w_all,Eigen::VectorXd &v_all,Eigen::MatrixXd &lamda_all,Eigen::VectorXd &m_all);
 int solveEquation(kltFeature f, int r,float K, float *dx, float *dy,float small,bool trackingMode);
 
 //int trackingKnownRF(float x1, float y1, float *x2, float *y2,Mat img1,Mat gradx1,Mat grady1,Mat img2,Mat gradx2,Mat grady2);
